requirements.txt
################################
###### Core dependencies #######
################################
torch==1.10.0
ai2-tango[all]==0.3.3
allennlp==2.8.0
transformers>=4.1,<4.12
tqdm==4.62.2
